refactor: remove unused gogitdesc experiment
remove this unused package entirely from active codebase so it doesnt
confuse `go mod tidy`.

this was good progress on getting git describe functionality added
into go-git library, but ultimately supporting the full feature set
would be a fair amount of work and testing, and I don't want to inherit
the work of doing that for now even if it would lead to a potential PR
and feature addition that would be valuable to go-git.

In addition, go-git currently adds quite a bit of filesize weight to
the binary of an app, even if you aren't using most of its features
(some inefficient import paths?), so this keeps things slimmer anyhow.

Removing in this commit and via PR to make it clear where to find this
work again in future if we want to re-investigate this path.