Add LICENSE
git is GPLv2, libgit2 is GPLv2 with a linking exception. I should use something GPL to adhere to the community standards here, but need a linking exception to make this easy to use as a library as well. Therefore, I assume LGPLv3 is simplest. If you need something else, please contact me.